wip(poly build): build command for Package & Dependency Management tools without support for plugins or build hooks
