+++ components/polylith/building/__init__.py
@@ -1,8 +1,10 @@
-from polylith.building.core import (
-    cleanup,
-    copy_and_rewrite_bricks,
-    copy_bricks_as_is,
-    get_work_dir,
-)
+from polylith.building.core import cleanup, copy_and_rewrite_bricks, copy_bricks_as_is
+from polylith.building.paths import calculate_destination_dir, get_work_dir
 
-__all__ = ["cleanup", "copy_and_rewrite_bricks", "copy_bricks_as_is", "get_work_dir"]
+__all__ = [
+    "calculate_destination_dir",
+    "cleanup",
+    "copy_and_rewrite_bricks",
+    "copy_bricks_as_is",
+    "get_work_dir",
+]
